# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Green Way Motors API
  version: 0.3.0
  description: Implemented backend contract. GreenWay_Backend_Master_Spec_v1.md is the product and API roadmap.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/admin/uploads/presign:
    post:
      tags: [Uploads]
      operationId: adminPresignUpload
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/PresignUploadInput' } } }
      responses:
        '201': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/uploads/complete:
    post:
      tags: [Uploads]
      operationId: adminCompleteUpload
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/CompleteUploadInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/vehicles/{id}/media:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Uploads]
      operationId: adminListVehicleMedia
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Data' } }
    post:
      tags: [Uploads]
      operationId: adminAttachVehicleMedia
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MediaAttachInput' } } }
      responses:
        '201': { $ref: '#/components/responses/Data' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/vehicles/{id}/media/{mediaId}:
    parameters:
      - { $ref: '#/components/parameters/UuidId' }
      - { name: mediaId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Uploads]
      operationId: adminUpdateVehicleMedia
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MediaPatchInput' } } }
      responses: { '200': { $ref: '#/components/responses/Data' } }
    delete:
      tags: [Uploads]
      operationId: adminDeleteVehicleMedia
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Media detached. } }
  /api/v1/admin/vehicles/{id}/media/reorder:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    post:
      tags: [Uploads]
      operationId: adminReorderVehicleMedia
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/MediaReorderInput' } } }
      responses: { '200': { $ref: '#/components/responses/Data' } }
...
components:
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: greenway_refresh_token
  responses:
...
    Data:
      description: Permission-scoped data with request ID metadata.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DataEnvelope' }
...
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    DataEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: { type: object, additionalProperties: true } }
        meta: { $ref: '#/components/schemas/RequestMeta' }
...
    RequestMeta:
      type: object
      required: [requestId]
      properties: { requestId: { type: string } }
...
    PresignUploadInput:
      type: object
      additionalProperties: false
      required: [purpose, mimeType, sizeBytes]
      properties:
        purpose: { type: string, enum: [vehicle_media] }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp, video/mp4] }
        sizeBytes: { type: integer, minimum: 1, maximum: 104857600 }
    CompleteUploadInput:
      type: object
      additionalProperties: false
      required: [uploadIntentId]
      properties: { uploadIntentId: { type: string, format: uuid } }
    MediaAttachInput:
      type: object
      additionalProperties: false
      required: [uploadIntentId]
      properties:
        uploadIntentId: { type: string, format: uuid }
        isPrimary: { type: boolean }
        sortOrder: { type: integer }
    MediaPatchInput:
      type: object
      additionalProperties: false
      properties:
        isPrimary: { type: boolean }
        sortOrder: { type: integer }
    MediaReorderInput:
      type: object
      additionalProperties: false
      required: [mediaIds]
      properties:
        { mediaIds: { type: array, uniqueItems: true, items: { type: string, format: uuid } } }
...
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        requestId: { type: string }
...